pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.0",
    "ipython>=9.7.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "langchain[openai]>=1.0.3",
    "langgraph>=1.0.2",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
